Express Setup
Changed the overall file structure again, removing index.html and log.txt, moved documentation to a public folder which is now a static through express. Express is now handling all interactions and serves a text log for / and a list of movies for /movies. There is also a morgan log output to the console using standard morgan format. Error handling has been included as well.